+++ test/list_test.exs
@@ -3,7 +3,7 @@ defmodule A.ListTest do
 
   doctest A.List
 
-  test "A.List.repeatedly/2" do
+  test "A.List.repeat/2" do
     counter = :counters.new(1, [])
 
     fun = fn ->
@@ -11,10 +11,10 @@ defmodule A.ListTest do
       :counters.get(counter, 1)
     end
 
-    assert [] = A.List.repeatedly(fun, 0)
-    assert [1, 2, 3, 4, 5] = A.List.repeatedly(fun, 5)
-    assert [6, 7, 8, 9, 10, 11, 12, 13, 14, 15] = A.List.repeatedly(fun, 10)
+    assert [] = A.List.repeat(fun, 0)
+    assert [1, 2, 3, 4, 5] = A.List.repeat(fun, 5)
+    assert [6, 7, 8, 9, 10, 11, 12, 13, 14, 15] = A.List.repeat(fun, 10)
 
-    assert_raise FunctionClauseError, fn -> A.List.repeatedly(fun, -1) end
+    assert_raise FunctionClauseError, fn -> A.List.repeat(fun, -1) end
   end
 end